+++ Events/ShipShutdownEvent.cs
@@ -7,7 +7,7 @@ namespace EddiEvents
     public class ShipShutdownEvent : Event
     {
         public const string NAME = "Ship shutdown";
-        public const string DESCRIPTION = "Triggered when your ship's system are shutdown";
+        public const string DESCRIPTION = "Triggered when your ship's system is forced to shutdown";
         public const string SAMPLE = @"{ ""timestamp"":""2017-01-05T23:15:06Z"", ""event"":""SystemsShutdown"" }";
 
         public ShipShutdownEvent(DateTime timestamp) : base(timestamp, NAME)